+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - (Documentation) Update ArangoDeployment CR auto-generated docs
 - (Documentation) Update ArangoBackup and ArangoBackupPolicy CR auto-generated docs
 - (Bugfix) Fix missing Pod Status case in the RuntimeContainerImageUpdateAction
+- (Feature) Member Memory Reservation
 
 ## [1.2.34](https://github.com/arangodb/kube-arangodb/tree/1.2.34) (2023-10-16)
 - (Bugfix) Fix make manifests-crd-file command